Fix rounding up in abbreviate
Roblox recently implemented a change to `tostring()` which makes it much
more performant, but breaks the `round()` function.

From the release notes for version 509:
> tostring() now uses a new cross-platform number formatting
> implementation that produces precise, round-trippable, shortest decimal
> output and does this 10 times faster than our old implementation
> depending on the platform and the input number.

Because `tostring()` now always returns the correct shortest
representation of the input number, `round(1.005, 2, false)` would
return `1.00` instead of `1.01`. This exposes the flaw in the original
source code for our `numberToString` method where it would use
`math.floor` instead of `math.ceil` when rounding up.

Reference:
https://developer.roblox.com/en-us/resources/release-note/Release-Notes-for-509
https://devforum.roblox.com/t/rounding-function-broken-by-recent-roblox-update/1626727